pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "evacusim"
version = "0.1.0"
description = "Reusable LLM-driven evacuation simulation framework (JuPedSim + Concordia)"
requires-python = ">=3.10"
dependencies = [
    "jupedsim>=1.0.0",
    "numpy>=1.24.0",
    "shapely>=2.0.0",
    "networkx>=3.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "PyYAML>=6.0",
    "python-dotenv>=1.0.0",
    "azure-ai-inference>=1.0.0b9",
    "azure-core>=1.29.0",
    "aiohttp>=3.9.0",
    "rich>=13.0.0",
    "sentence-transformers>=2.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["evacusim*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
